macos/requirements.txt
rumps
pyperclip
requests
python-dotenv
py2app
setuptools==70.3.0
